requirements.txt
discord.py
youtube-dl
python-dotenv